Switch to a dimension-based caching system for generators.
I have no idea how efficient this is, but it's actually now allowing
Superflat Caves to generate Cave Generator caves outside of the initial
spawn chunks.

I don't know what impact it will have on anything else; presumably none.

Each of the changes should, in theory, have no impact on the rest of the
code, but I'm not super-familiar with it.

Indeed, whenever the generators are fetched the WeakReferences are
updated.

I'm not sure whether this will reduce performance or not.